build(ui): create prebuild script to load env
